Change executor image back to alpine
The problem generated by using alpine seems to be due to incompatible
executables between machines, so if you compiled on your machine then
sent to the docker, everything would fail, but if you actually send a
source code to the compiler running on docker and save the output as an
output file and then send that to the executor, it works just fine with
an alpine image.

Also, gcc and g++ libraries are needed to run those binaries, so I'm
installing them, and while at it, I'm installing python3 too.

Signed-off-by: Pedro Sousa <pedroteosousa@gmail.com>