Consume submission queue on orchestrator route
This commit is a very crude closing of the queue loop. The intended
behaviour is for the orchestrator to just consume the queue and send
information to verdict, but there are a few problems when a connection
with the queue might close for some reason and it is not clear for me
what would be the best course of action to handle these cases.

For now, the orchestrator has a new route, /run, that consumes the queue
and sends the submissions to verdict. This request of course should not
exist, this is just a the very basic functionality so we can work on top
of it.