Use reflect instead of json on graphqlMock
We were calling Marshal then Unmarshal to transfer an interface into a
pointer to an interface, but we can do this directly using reflect.

This will panic if there is a problem with the types, but we will only
use this in tests. This keeps graphqlMock regeric enough to be used on
other server tests.