Move graphqlMock to utils/graphql.go
This graphql mock will be used by other server tests, so we are moving
it to a common place. Maybe it's best to create a utils just for tests
and put this there.

Also breaking down submission type into all its components, and calling
it Info, so we don't have to redeclared its types.