Fix executor server tests
Executor tests were not yet using GraphqlMock, this is just an
adaptation to handle this. It's important to note that executor no
longer received an executable on the request body, instead looking for
an executable on /var, so we are saving the executable object directly
into /var before calling the server.