Fix travis.yml to work with tests on docker
Now the test.sh creates the coverage file inside a coverage folder,
which is a volume shared with the host machine, also inside a coverage
folder. After running the docker, this file is copied from the coverage
folder to ./coverage.txt so the coverage script can run as normally.

We run gofmt twice, because travis needs to know if this command fails,
but it won't know this from docker, since only the last command from
test.sh will determine the docker exit status, which, is this case, is
the `go test`, not gofmt.