Use kubernetes client-go on orchestrator
Orchestrator tries to create the mart pod when it reads a submission
from the queue. mart pod is no longer at k8s.
The pod fails to be created because orchestrator does not have the
required authorization to do so.

Kill port forward now uses `fuser` instead of `kill`.

Created a script to build docker images and push them to docker hub.