Remove server from verdict service
Instead of waiting for a POST with the submission ID, the verdict now gets the ID from the environment variable SUBMISSION_ID. For that reason we were able to remove its server.

Needed to add the ClusterRoleBinding to provide admin-access to the Orchestrator service, so it can create another pod on the cluster.

Now the Orchestrator sets the 'Pending' state for every submission it dequeues from RabbitMQ.

Added 3 retries to MakeSubmissionRequest, since the verdict service may start creating requests even before the compiler server starts.