Use exponential backoff on verdict calls
Using 7 retries of power of two delays when verdict calls compiler and
executor, this adds up to about 2 minutes of retries. In some tests the
previous linear wait times were not enough in some cases, and the
verdict would exit without being able to send the request to compile the
source.

The amount of time compiler and executor keep listening before giving up
was increased to 2 minutes to reflect the change in the retry logic,
since now verdict can spend a bit more than a minute between the last
two retries.